Moved FileTag class, new colors
