+++ Files/UserControls/MenuFlyoutItemFileTag.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Files.Filesystem;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Windows.UI;
@@ -34,24 +35,6 @@ namespace Files.UserControls
         {
             this.InitializeComponent();
             this.ItemsSource = new List<FileTag>();
-        }
-    }
-
-    public class FileTag
-    {
-        public string Tag { get; set; }
-        public SolidColorBrush Color { get; set; }
-        
-        public FileTag(string tag = null, SolidColorBrush color = null)
-        {
-            Tag = tag;
-            Color = color ?? new SolidColorBrush(Colors.Transparent);
-        }
-
-        public FileTag(string tag = null, Color? color = null)
-        {
-            Tag = tag;
-            Color = new SolidColorBrush(color ?? Colors.Transparent);
         }
     }
 }
